GCP Batch - reorganize directories (tests)
